„scripts/start-server.sh“ ändern

+++ scripts/start-server.sh
@@ -1,16 +1,15 @@
 #!/bin/bash
-sleep infintiy
-
-export NWNX_DOTNET_SKIP=y
-export LD_PRELOAD="./NWNX_Core.so"
-export NWNX_METRICS_INFLUXDB_HOST=127.0.0.1
-export NWNX_METRICS_INFLUXDB_PORT=3306
-export NWNX_REDIS_HOST=127.0.0.1
+export NWNX_SQL_TYPE=mysql
+export NWNX_SQL_HOST=localhost
 export NWNX_SQL_USERNAME=nwnee
 export NWNX_SQL_PASSWORD=nwnee
 export NWNX_SQL_DATABASE=nwnee
+export NWNX_SQL_QUERY_METRICS=true
+export NWNX_CORE_LOG_LEVEL=${LOG_LVL}
+export NWNX_CORE_LOAD_PATH=/nwnx/binaries
+
 CUR_V="$(find ${SERVER_DIR} -name build* | cut -d 'd' -f2)"
-CUR_V_BIN="$(find ${SERVER_DIR} -name binaries* | cut -d 's' -f2)"
+CUR_V_BIN="$(find ${SERVER_DIR} -name binariesbuild* | cut -d 'd' -f2)"
 LAT_V="$(curl -s https://api.github.com/repos/nwnxee/unified/releases/latest | grep tag_name | cut -d '"' -f4 | cut -d 'd' -f2)"
 if [ "${NWNEE_V}" == "latest" ]; then
 	NWNEE_V=$LAT_V
@@ -37,23 +36,68 @@ if [ -z "$CUR_V" ]; then
 	fi
 	unzip -o ${SERVER_DIR}/nwnee-dedicated-${NWNEE_V}.zip
 	touch ${SERVER_DIR}/build${NWNEE_V}
-    rm ${SERVER_DIR}/nwnee-dedicated-${NWNEE_V}.zip
+	rm ${SERVER_DIR}/nwnee-dedicated-${NWNEE_V}.zip
 fi
 
 if [ -z "$CUR_V_BIN" ]; then
 	echo "---NWN:EE Binaries not found, installing v${NWNEE_V}!---"
-	cd ${SERVER_DIR}/bin/linux-x86
+	if [ ! -d ${SERVER_DIR}/binaries ]; then
+		mkdir ${SERVER_DIR}/binaries
+	fi
+	cd ${SERVER_DIR}/binaries
 	if wget https://github.com/nwnxee/unified/releases/download/build${NWNEE_V}/NWNX-EE.zip ; then
+		echo "---Sucessfully downloaded NWN:EE Binaries---"
+	else
+		echo "---Something went wrong, can't download NWN:EE Binaries, putting server in sleep mode---"
+		sleep infinity
+	fi
+	unzip -o ${SERVER_DIR}/binaries/NWNX-EE.zip
+	mv ${SERVER_DIR}/binaries/Binaries/* ${SERVER_DIR}/binaries
+	rm -R ${SERVER_DIR}/binaries/Binaries
+	touch ${SERVER_DIR}/binariesbuild${NWNEE_V}
+	rm ${SERVER_DIR}/binaries/NWNX-EE.zip
+fi
+
+echo "---Server Version Check---"
+if [ "${NWNEE_V}" != "$CUR_V" ]; then
+	echo "---Version missmatch v$CUR_V installed, installing v${NWNEE_V}---"
+    rm ${SERVER_DIR}/build$CUR_V
+    rm -R ${SERVER_DIR}/bin
+    rm -R ${SERVER_DIR}/data
+	cd ${SERVER_DIR}
+	if wget https://nwnx.io/nwnee-dedicated-${NWNEE_V}.zip ; then
 		echo "---Sucessfully downloaded NWN:EE Dedicated Server---"
 	else
 		echo "---Something went wrong, can't download NWN:EE Dedicated Server, putting server in sleep mode---"
 		sleep infinity
 	fi
-	unzip -o ${SERVER_DIR}/bin/linux-x86/NWNX-EE.zip
-    mv ${SERVER_DIR}/bin/linux-x86/Binaries/* ${SERVER_DIR}/bin/linux-x86
-    rm -R ${SERVER_DIR}/bin/linux-x86/Binaries
-	touch ${SERVER_DIR}/binaries${NWNEE_V}
-    rm ${SERVER_DIR}/bin/linux-x86/NWNX-EE.zip
+	unzip -o ${SERVER_DIR}/nwnee-dedicated-${NWNEE_V}.zip
+	touch ${SERVER_DIR}/build${NWNEE_V}
+	rm ${SERVER_DIR}/nwnee-dedicated-${NWNEE_V}.zip
+elif [ "${NWNEE_V}" == "$CUR_V" ]; then
+	echo "---Server versions match! Installed: v$CUR_V | Preferred: v${NWNEE_V}---"
+fi
+
+echo "---Binaries Version Check---"
+if [ "${NWNEE_V}" != "$CUR_V_BIN" ]; then
+	echo "---Version missmatch v$CUR_V_BIN installed, installing v${NWNEE_V}---"
+    rm ${SERVER_DIR}/binariesbuild$CUR_V_BIN
+    rm -R ${SERVER_DIR}/binaries
+    mkdir ${SERVER_DIR}/binaries
+	cd ${SERVER_DIR}/binaries
+	if wget https://github.com/nwnxee/unified/releases/download/build${NWNEE_V}/NWNX-EE.zip ; then
+		echo "---Sucessfully downloaded NWN:EE Binaries---"
+	else
+		echo "---Something went wrong, can't download NWN:EE Binaries, putting server in sleep mode---"
+		sleep infinity
+	fi
+	unzip -o ${SERVER_DIR}/binaries/NWNX-EE.zip
+	mv ${SERVER_DIR}/binaries/Binaries/* ${SERVER_DIR}/binaries
+	rm -R ${SERVER_DIR}/binaries/Binaries
+	touch ${SERVER_DIR}/binariesbuild${NWNEE_V}
+	rm ${SERVER_DIR}/binaries/NWNX-EE.zip
+elif [ "${NWNEE_V}" == "$CUR_V_BIN" ]; then
+	echo "---Binaries versions match! Installed: v$CUR_V_BIN | Preferred: v${NWNEE_V}---"
 fi
 
 echo "---Prepare Server---"
@@ -64,6 +108,15 @@ fi
 if [ ! -d ${SERVER_DIR}/user ]; then
 	mkdir ${SERVER_DIR}/user
 fi
+if [ ! -d "${SERVER_DIR}/Neverwinter Nights" ]; then
+	mkdir "${SERVER_DIR}/Neverwinter Nights"
+fi
+if [ ! -d "${SERVER_DIR}/Neverwinter Nights/modules" ]; then
+	mkdir "${SERVER_DIR}/Neverwinter Nights/modules"
+fi
+if [ ! -d "${SERVER_DIR}/Neverwinter Nights/hak" ]; then
+	mkdir "${SERVER_DIR}/Neverwinter Nights/hak"
+fi
 echo "---Configuring Redis---"
 sleep 5
 echo "CONFIG SET dir ${SERVER_DIR}/redis" | redis-cli
@@ -72,9 +125,17 @@ echo "BGSAVE" | redis-cli
 echo "---Checking for old logs---"
 find ${SERVER_DIR} -name "MariaDBLog.0" -exec rm -f {} \;
 find ${SERVER_DIR} -name "RedisLog.0" -exec rm -f {} \;
-
 chmod -R 777 ${SERVER_DIR}
 
-echo "---Start Server---"
+if [ ! "$(ls -A "${SERVER_DIR}/Neverwinter Nights/modules")" ]; then
+	echo "-----------------------------------------------------------------"
+	echo "---------Your 'modules' folder is empty, please put your---------"
+	echo "---required 'module' files in the 'Neverwinter Nights/modules'---"
+	echo "------------folder and restart the container, putting------------"
+    echo "---------------------server into sleep mode----------------------"
+	echo "-----------------------------------------------------------------"
+	sleep infinity
+fi
+
 cd ${SERVER_DIR}/bin/linux-x86
-${SERVER_DIR}/bin/linux-x86/nwserver-linux -maxclients ${MAX_CLIENTS} -minlevel ${MINLEVEL} -maxlevel ${MAXLEVEL} -pauseandplay ${PAUSEAPLAY} -pvp ${PVP} -servervault ${SERVERVAULT} -elc ${ELC} -ilr ${ILR} -oneparty ${ONEPARTY} -difficulty ${DIFF} -autosaveinterval ${AUTO_SAV_I} -playerpassword ${PPW} -adminpassword ${APWD} -servername ${SRV_NAME} -publicserver ${PUBLIC_SRV} -reloadwhenempty ${RLD_W_E} -port ${GAME_PORT} -userdirectory ${SERVER_DIR}/user
+LD_PRELOAD=/nwnee/binaries/NWNX_Core.so ${SERVER_DIR}/bin/linux-x86/nwserver-linux -module "${MOD_NAME}" -maxclients ${MAX_CLIENTS} -minlevel ${MINLEVEL} -maxlevel ${MAXLEVEL} -pauseandplay ${PAUSEAPLAY} -pvp ${PVP} -servervault ${SERVERVAULT} -elc ${ELC} -ilr ${ILR} -oneparty ${ONEPARTY} -difficulty ${DIFF} -autosaveinterval ${AUTO_SAV_I} -playerpassword ${PPW} -adminpassword ${APWD} -servername "${SRV_NAME}" -publicserver ${PUBLIC_SRV} -reloadwhenempty ${RLD_W_E} -port ${GAME_PORT} -userdirectory "${SERVER_DIR}/Neverwinter Nights"
